moved extra WR stuff to extra
