--- lib/Extension/WorseReflection/GotoDefinition/Exception/GotoDefinitionException.php
@@ -1,7 +1,0 @@
-<?php
-
-namespace Phpactor\Extension\WorseReflection\GotoDefinition\Exception;
-
-class GotoDefinitionException extends \DomainException
-{
-}
