+++ lib/Extension/WorseReflectionExtra/GotoDefinition/Exception/GotoDefinitionException.php
@@ -1,0 +1,7 @@
+<?php
+
+namespace Phpactor\Extension\WorseReflectionExtra\GotoDefinition\Exception;
+
+class GotoDefinitionException extends \DomainException
+{
+}
