+++ tests/Unit/Extension/Core/GotoDefinition/GotoDefinitionTest.php
@@ -6,7 +6,7 @@ use PHPUnit\Framework\TestCase;
 use Phpactor\WorseReflection\Core\Name;
 use Phpactor\WorseReflection\Core\Reflection\ReflectionFunction;
 use Phpactor\WorseReflection\Reflector;
-use Phpactor\Extension\WorseReflection\GotoDefinition\GotoDefinition;
+use Phpactor\Extension\WorseReflectionExtra\GotoDefinition\GotoDefinition;
 use Phpactor\WorseReflection\Core\Inference\Symbol;
 use Phpactor\WorseReflection\Core\Inference\SymbolContext;
 use Phpactor\WorseReflection\Core\Position;
@@ -17,12 +17,12 @@ use Phpactor\WorseReflection\Core\Reflection\ReflectionMethod;
 use Phpactor\WorseReflection\Core\ClassName;
 use Phpactor\WorseReflection\Core\Exception\NotFound;
 use Phpactor\WorseReflection\Core\SourceCode;
-use Phpactor\Extension\WorseReflection\GotoDefinition\GotoDefinitionResult;
+use Phpactor\Extension\WorseReflectionExtra\GotoDefinition\GotoDefinitionResult;
 use Phpactor\WorseReflection\Core\Reflection\ReflectionConstant;
 use Phpactor\WorseReflection\Core\Reflection\Collection\ReflectionConstantCollection;
 use Phpactor\WorseReflection\Core\Reflection\ReflectionProperty;
 use Phpactor\WorseReflection\Core\Reflection\Collection\ReflectionPropertyCollection;
-use Phpactor\Extension\WorseReflection\GotoDefinition\Exception\GotoDefinitionException;
+use Phpactor\Extension\WorseReflectionExtra\GotoDefinition\Exception\GotoDefinitionException;
 
 class GotoDefinitionTest extends TestCase
 {
