+++ tests/Unit/Extension/WorseReflection/Rpc/GotoDefinitionHandlerTest.php
@@ -6,7 +6,7 @@ use Phpactor\Extension\Rpc\Handler;
 use Phpactor\WorseReflection\Core\SourceCode;
 use Phpactor\WorseReflection\ReflectorBuilder;
 use Phpactor\Tests\Unit\Extension\Rpc\Handler\HandlerTestCase;
-use Phpactor\Extension\WorseReflection\Rpc\GotoDefinitionHandler;
+use Phpactor\Extension\WorseReflectionExtra\Rpc\GotoDefinitionHandler;
 
 class GotoDefinitionHandlerTest extends HandlerTestCase
 {
